+++ src/services/member/utils/index.ts
@@ -1,3 +1,3 @@
 export * from './role';
-export * from './nickname';
+export * from './createRandomNickname';
 export * from './types';
